Sword overhaul
- Created a separate shop for swords
- Swords are now pre-determined rather than mathematically generated
- Fixed a bug with damage indicator for durability

+++ plugin/src/main/java/me/theguyhere/villagerdefense/plugin/game/models/items/VDItem.java
@@ -83,7 +83,7 @@ public abstract class VDItem {
         meta.setLore(lores);
 
         // Set damage indicator
-        damage.setDamage((int) (((item.getType().getMaxDurability() - durability.get()) * 1. / maxDur.get()) *
+        damage.setDamage((int) item.getType().getMaxDurability() - (int) (durability.get() * 1. / maxDur.get() *
                 item.getType().getMaxDurability()));
 
         item.setItemMeta(meta);
